Fixed reset packet bugs; updated flags
If LibSSH encountered a reset packet immediately after a connect scan,
it would crash the program. In cases with Samba deception, an exception
would be thrown and ignored. We catch and alert on the exception as a
reset packet being sent is still a form of interaction that should be
logged. Software no longer crashes in cases with LibSSH deception.

In the previous version, when a user used any flag, all deception was
disabled unless the user explicitly used the --smbD or --sshD flags to
re-enable deception. This is counterintuitive. Now, --smbD and --sshD
disable deception, instead. The options -a and --all are removed
since they are now the default.